+++ modules/sofa/component/collision/CapsuleModel.h
@@ -118,7 +118,7 @@ public:
 
     virtual void resize(int size);
 
-    virtual void computeBoundingTree(int maxDepth=0);
+    virtual void computeBoundingTree(size_t maxDepth=0);
 
     //virtual void computeContinuousBoundingTree(SReal dt, int maxDepth=0);
 
@@ -179,7 +179,7 @@ public:
 
     Data<VecReal > & writeRadii();
 protected:
-    core::behavior::MechanicalState<DataTypes>* _mstate;    
+    core::behavior::MechanicalState<DataTypes>* _mstate;
 };
 
 template<class DataTypes>
