+++ modules/sofa/component/collision/OBBModel.h
@@ -166,7 +166,7 @@ public:
 
     virtual void resize(int size);
 
-    virtual void computeBoundingTree(int maxDepth=0);
+    virtual void computeBoundingTree(size_t maxDepth=0);
 
     //virtual void computeContinuousBoundingTree(SReal dt, int maxDepth=0);
